corrected bug with the attacker getting 'knocked down' but displaying that the target was knocked down instead
