+++ jest
@@ -1,3 +1,3 @@
 #!/usr/bin/env zsh
 
-node --experimental-vm-modules node_modules/.bin/jest
+node --experimental-vm-modules node_modules/.bin/jest $1
